requirements.txt
aiogram==3.10.0
google-generativeai==0.5.5
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.3
pytz==2024.1